タスク削除処理を削除フラグ考慮形式に修正

+++ src/main/java/com/portfolio/taskapp/MyTaskManager/task/service/TaskService.java
@@ -122,6 +122,10 @@ public class TaskService {
 
   @Transactional
   public void deleteTask(String taskPublicId) {
+    // 削除対象のnullチェック
+    if (repository.findTaskIdByTaskPublicId(taskPublicId) == null) {
+      throw new RecordNotFoundException("task not found");
+    }
     repository.deleteTask(taskPublicId);
   }
 }
